backend/requirements.txt
django>=3.2,<4
django-environ<1
djangorestframework<3.14
django-storages>=1.13.2,<2
django-filter>=23.2,<24
django-prometheus==2.2.0
celery>=5.2,<5.3
django-celery-beat<3
aioredis<3
django-redis<6
boto3<2
psycopg2-binary<3
gunicorn<21
uvicorn<1
sentry-sdk>=1.22,<2
uvloop>=0.17,<1
Pillow>=9.0.1
json-logging>=1.3,<2
isort[pyproject]<5
flake8>=5
flake8-isort>=3,<4
flake8-commas>=2.1,<3
pytest>=7,<8
mixer>=7.2,<8
coverage>=6.3,<7
PyJWT[crypto]==2.7.0
Faker>=8,<9
httpx<1
pytest-django>=4.5.2,<4.6
pytest-sugar==0.9.4
ipython==8.15.0+fix
cryptography==41.0.1
django-mptt>=0.13.4,<1
django-currentuser==0.5.3
raven==6.10.0
debugpy<2
django-mdeditor==0.1.20
django-mptt-admin==2.0.0
django-debug-toolbar==3.2.4
django-cleanup>=6,<7